# Conflicts:
#	erp-common/erp-common-message/src/main/java/com/common/message/constant/RocketMqNewConsumerGroup.java
#	erp-common/erp-common-message/src/main/java/com/common/message/constant/RocketMqNewTopic.java
#	erp-rpc/erp-rpc-sys/src/main/java/com/erp/rpc/sys/feign/SysUserFeign.java
#	erp-server/erp-server-dmp/src/main/java/com/erp/server/dmp/inout/job/DmpInputTaskJob.java

+++ erp-sdk/erp-sdk-oms-wildberries/src/main/java/com/sdk/oms/wildberries/dto/OrderRequest.java
@@ -14,7 +14,7 @@ import lombok.Data;
 @Builder
 public class OrderRequest {
     private Integer limit;
-    private Integer next;
+    private Long next;
     //时间戳
     private Long dateFrom;
     private Long dateTo;
